Update WeatherData.swift
Temperature isn't Int service now returns Double

+++ Project 25 - WeatherExtension/WeatherDemo/WeatherData.swift
@@ -10,7 +10,7 @@ import Foundation
 
 open class WeatherData: NSObject {
     
-    open var temperature:Int = 0
-    open var weather:String = ""
+ open var temperature:Double = 0.0
+ open var weather:String = ""
     
 }
